Encapsulated route handlers for better testing.
